requirements.txt
absl-py==2.1.0
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
apprise==1.9.2
asgi-lifespan==2.1.0
astunparse==1.6.3
asyncpg==0.30.0
attrs==25.1.0
bleach==6.2.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
coolname==2.2.0
cryptography==44.0.1
cycler==0.12.1
dateparser==1.2.1
Deprecated==1.2.18
docker==7.1.0
duckdb==1.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.7
grpcio==1.70.0
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.1
hyperframe==6.1.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
Jinja2==3.1.5
jinja2-humanize-extension==0.4.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
oauthlib==3.2.2
opentelemetry-api==1.30.0
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==3.0.0
pillow==11.1.0
polars==1.22.0
prefect==3.2.6
prometheus_client==0.21.1
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.1
pywin32==308
PyYAML==6.0.2
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-addons==0.22.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
tensorflow_keras==0.1
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
toml==0.10.2
tqdm==4.67.1
typeguard==2.13.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
websockets==13.1
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
zipp==3.21.0
